Use tile coordinates for moving and the server, and map scrolling Load library map from the server and grab princess and levers info from there

+++ mmo/src/mmo/protocol/dto.kt
@@ -25,7 +25,7 @@ data class ClientSetLang(val lang: String) : ClientPacket
 data class ClientSay(val text: String) : ClientPacket
 
 @Serializable
-data class ClientRequestMove(val x: Double, val y: Double) : ClientPacket
+data class ClientRequestMove(val x: Int, val y: Int) : ClientPacket
 
 @Serializable
 data class ClientRequestInteract(val entityId: Long) : ClientPacket
